move back transform and flatten functions
