+++ static/app/views/dashboardsV2/datasetConfig/errorsAndTransactions.tsx
@@ -1,4 +1,3 @@
-import omit from 'lodash/omit';
 import trimStart from 'lodash/trimStart';
 
 import {doEventsRequest} from 'sentry/actionCreators/events';
@@ -20,10 +19,8 @@ import {CustomMeasurementCollection} from 'sentry/utils/customMeasurements/custo
 import {EventsTableData, TableData} from 'sentry/utils/discover/discoverQuery';
 import {MetaType} from 'sentry/utils/discover/eventView';
 import {
-  DURATION_UNITS,
   getFieldRenderer,
   RenderFunctionBaggage,
-  SIZE_UNITS,
 } from 'sentry/utils/discover/fieldRenderers';
 import {
   errorsAndTransactionsAggregateFunctionOutputType,
@@ -61,6 +58,10 @@ import {
 } from '../utils';
 import {EventsSearchBar} from '../widgetBuilder/buildSteps/filterResultsStep/eventsSearchBar';
 import {CUSTOM_EQUATION_VALUE} from '../widgetBuilder/buildSteps/sortByStep';
+import {
+  flattenMultiSeriesDataWithGrouping,
+  transformSeries,
+} from '../widgetCard/widgetQueries';
 
 import {DatasetConfig, handleOrderByReset} from './base';
 
@@ -332,65 +333,6 @@ function filterYAxisOptions(displayType: DisplayType) {
 
     return option.value.kind === FieldValueKind.FUNCTION;
   };
-}
-
-function transformSeries(stats: EventsStats, seriesName: string, field: string): Series {
-  const unit = stats.meta?.units?.[getAggregateAlias(field)];
-  // Scale series values to milliseconds or bytes depending on units from meta
-  const scale = (unit && (DURATION_UNITS[unit] ?? SIZE_UNITS[unit])) ?? 1;
-  return {
-    seriesName,
-    data:
-      stats?.data?.map(([timestamp, counts]) => {
-        return {
-          name: timestamp * 1000,
-          value: counts.reduce((acc, {count}) => acc + count, 0) * scale,
-        };
-      }) ?? [],
-  };
-}
-
-/**
- * Multiseries data with a grouping needs to be "flattened" because the aggregate data
- * are stored under the group names. These names need to be combined with the aggregate
- * names to show a series.
- *
- * e.g. count() and count_unique() grouped by environment
- * {
- *    "local": {
- *      "count()": {...},
- *      "count_unique()": {...}
- *    },
- *    "prod": {
- *      "count()": {...},
- *      "count_unique()": {...}
- *    }
- * }
- */
-function flattenMultiSeriesDataWithGrouping(
-  result: EventsStats | MultiSeriesEventsStats,
-  queryAlias: string
-): SeriesWithOrdering[] {
-  const seriesWithOrdering: SeriesWithOrdering[] = [];
-  const groupNames = Object.keys(result);
-
-  groupNames.forEach(groupName => {
-    // Each group contains an order key which we should ignore
-    const aggregateNames = Object.keys(omit(result[groupName], 'order'));
-
-    aggregateNames.forEach(aggregate => {
-      const seriesName = `${groupName} : ${aggregate}`;
-      const prefixedName = queryAlias ? `${queryAlias} > ${seriesName}` : seriesName;
-      const seriesData: EventsStats = result[groupName][aggregate];
-
-      seriesWithOrdering.push([
-        result[groupName].order || 0,
-        transformSeries(seriesData, prefixedName, seriesName),
-      ]);
-    });
-  });
-
-  return seriesWithOrdering;
 }
 
 function transformEventsResponseToSeries(
